+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-[jdbi3]
+* [jdbi3]
     - Fix unit of work closing after connection error (preventing application recover) (#35)
+* [jdbi2]
+    - Fix unit of work closing after connection error (preventing application recover) (#35)    
 
 ### 5.1.0-1 (2020-06-02)
 * [gsp]